Validations des entrées lors de la requête put + adaptations des middleware pour éviter la suppression de l'image

+++ backend/middlewares/deleteImage.js
@@ -3,29 +3,38 @@ const fs = require("fs")
 const Sauce = require('../models/sauceModel');
 
 module.exports = (req, res, next) => {
-    Sauce.findOne({_id:req.params.id})
-    .then(sauce => {
+
+    // IF UPDATE INFO : NO DELETE
+    if(req.headers['content-type'] == "application/json"){
+         next()
+    }
         
-        let filename = sauce.imageUrl;
-        filename = filename.split('/')
-        filename = filename[filename.length - 1]
+    // IF INFO AND IMAGE : DELETE
+    else {
+        Sauce.findOne({_id:req.params.id})
+        .then(sauce => {
+            
+            let filename = sauce.imageUrl;
+            filename = filename.split('/')
+            filename = filename[filename.length - 1]
 
-        let pathDelete = path.join(process.cwd(),'images/', filename)
-        console.log(pathDelete);
+            let pathDelete = path.join(process.cwd(),'images/', filename)
+            console.log(pathDelete);
 
-        if (sauce.userId == req.auth.userId){
-            fs.unlink(pathDelete, () => {
-                console.log("Suppression avant update");
-            })
-            next()
-        }
+            if (sauce.userId == req.auth.userId){
+                fs.unlink(pathDelete, () => {
+                    console.log("Suppression depuis le middleware");
+                })
+                next()
+            }
 
-        else {
-            res.status(403).json({message : "Utilisateur non autorisé"})
-            console.log("non autorisé");
-        }
+            else {
+                res.status(403).json({message : "Utilisateur non autorisé"})
+                console.log("non autorisé");
+            }
 
-    })
+        })
 
-    .catch(err => console.log(err))
+        .catch(err => console.log(err))
+    }
 };
